line and pixel drawing works again
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@2995 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/video/canvas/glide2common/glide2common2d.h
@@ -93,7 +93,7 @@ public:
   //  Glide-specific procedures
   void SetTMUPalette(int tmu);
   /// Draw a line
-  virtual void DrawLine( int x1, int y1, int x2, int y2, int color);
+  virtual void DrawLine( float x1, float y1, float x2, float y2, int color);
   /// Draw a pixel
   virtual void DrawPixel (int x, int y, int color);
   /// Write a single character
